refactoring: separated blocks utilities as blocks.py
